// src/lib/spacetime/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const AppSetting = __t.object("AppSetting", {
  id: __t.u8(),
  timeLimitSeconds: __t.u32(),
  verbsPerLevel: __t.u32(),
  activeVerbCategory: __t.u32(),
  updatedAt: __t.timestamp(),
  updatedBy: __t.identity(),
});
export type AppSetting = __Infer<typeof AppSetting>;

export const GameSession = __t.object("GameSession", {
  id: __t.u64(),
  userIdentity: __t.identity(),
  levelAttempted: __t.u32(),
  errorsCount: __t.u32(),
  durationSeconds: __t.u32(),
  completedAt: __t.timestamp(),
  isPerfectRun: __t.bool(),
  clientTimestampStart: __t.timestamp(),
  clientTimestampEnd: __t.timestamp(),
});
export type GameSession = __Infer<typeof GameSession>;

export const LevelRunDetail = __t.object("LevelRunDetail", {
  durationSeconds: __t.u32(),
  isPerfectRun: __t.bool(),
  completedAt: __t.timestamp(),
  errorsCount: __t.u32(),
  clientTimestampStart: __t.timestamp(),
});
export type LevelRunDetail = __Infer<typeof LevelRunDetail>;

export const LevelRunsResponse = __t.object("LevelRunsResponse", {
  get runs() {
    return __t.array(LevelRunDetail);
  },
  get verbs() {
    return __t.array(LevelVerbDetail);
  },
});
export type LevelRunsResponse = __Infer<typeof LevelRunsResponse>;

export const LevelVerbDetail = __t.object("LevelVerbDetail", {
  id: __t.u64(),
  infinitive: __t.string(),
  pastSimple: __t.string(),
  pastParticiple: __t.string(),
});
export type LevelVerbDetail = __Infer<typeof LevelVerbDetail>;

export const User = __t.object("User", {
  identity: __t.identity(),
  usernameLower: __t.string(),
  username: __t.string(),
  passwordHash: __t.string(),
  role: __t.string(),
  currentLevelCap: __t.u32(),
  createdAt: __t.timestamp(),
});
export type User = __Infer<typeof User>;

export const UserAttempt = __t.object("UserAttempt", {
  identity: __t.identity(),
  lastStatus: __t.string(),
  lastNewLevel: __t.u32(),
  lastCompletedAt: __t.timestamp(),
});
export type UserAttempt = __Infer<typeof UserAttempt>;

export const UserLevelDetail = __t.object("UserLevelDetail", {
  levelAttempted: __t.u32(),
  totalRuns: __t.u64(),
  perfectRuns: __t.u64(),
  bestTimeSeconds: __t.option(__t.u32()),
  globalRank: __t.option(__t.u64()),
});
export type UserLevelDetail = __Infer<typeof UserLevelDetail>;

export const UserOverview = __t.object("UserOverview", {
  userIdentity: __t.string(),
  username: __t.string(),
  currentLevelCap: __t.u32(),
  totalRuns: __t.u64(),
  totalPerfectRuns: __t.u64(),
});
export type UserOverview = __Infer<typeof UserOverview>;

export const Verb = __t.object("Verb", {
  id: __t.u64(),
  levelGroup: __t.u32(),
  category: __t.u32(),
  infinitive: __t.string(),
  pastSimple: __t.string(),
  pastParticiple: __t.string(),
  active: __t.bool(),
});
export type Verb = __Infer<typeof Verb>;